composition calculations independent of Configuration
